Separete de logic of UI, and some code commented
