+++ src/App.jsx
@@ -1,34 +1,18 @@
-import React , { useEffect, useState, createContext } from 'react';
+import React , { createContext } from 'react';
 import { NavBar } from './Components/NavBar/NavBar.jsx';
 import { Pagination } from './Components/Pagination/Pagination.jsx';
 import { Characters } from './Components/Characters/Characters.jsx';
+import { AppLogic } from './AppLogic';
 import './App.css';
 
 const AppContext = createContext(null);
 
 function App() {
-  const [ characters , setCharacters ] = useState([]);
-  const [ info  , setInfo ] = useState({});
+  //Desestructuring the states and functions to send as value on AppContext
+  const { characters , info , onPrev , onNext } = AppLogic();
 
-  const fetchData = url => {
-    fetch( url )
-    .then( response => response.json() )
-    .then( data => {
-       setCharacters(data.results)
-        setInfo(data.info)  
-    })
-    .catch( error => console.log(error) )
-  }
-
-  useEffect( ()=>{
-    fetchData('https://rickandmortyapi.com/api/character');
-  }, [] )
-
-  const onPrev = () => fetchData(info.prev);
-  const onNext = () => fetchData(info.next);
-  
   return (
-    <AppContext.Provider className="App" value={{
+    <AppContext.Provider value={{
       onPrev , onNext , characters , info
     }}>
       <NavBar/>
@@ -37,6 +21,7 @@ function App() {
       <Pagination />
     </AppContext.Provider>
   );
+
 }
 
 export { App , AppContext };
